# api/openapi.yaml
...
info:
  description: REST/JSON transport for the qoltanba digital-signature service (RK
    / Kalkan). One call handles one item. Binary fields are base64-encoded strings;
    times are RFC 3339. All request and response keys are lowerCamelCase. Generated
    from the Go types — do not edit by hand. The proprietary Kalkan library is bring-your-own
    (BYOL) and is not part of this service.
  title: qoltanba REST API
  version: v1
openapi: 3.1.0
paths:
...
  /verify/portal:
    get:
      operationId: verifyPortal
      responses:
        "200":
          content:
            text/html:
              schema:
                type: string
          description: The upload page
      summary: Upload form for checking a signature (off by default — it accepts uploads
        from anyone who can reach it)
      tags:
      - portal
    post:
      operationId: verifyPortalSubmit
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                document:
                  description: the original file — only a detached signature needs
                    it
                  format: binary
                  type: string
                signature:
                  description: the signed container (.p7s, .cms, .xml); format is
                    detected from the bytes
                  format: binary
                  type: string
              required:
              - signature
              type: object
        required: true
      responses:
        "200":
          content:
            text/html:
              schema:
                type: string
          description: The verification card for the uploaded container
        "400":
          content:
            text/html:
              schema:
                type: string
          description: Nothing to check, or the check could not be run — the page
            says which
      summary: Check an uploaded container and render the verification card
      tags:
      - portal
...
servers:
- description: Default REST work port (-http :8080)
  url: http://localhost:8080
...